django_backend/requirements.txt
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
djangorestframework==3.14.0
mysql-connector-python==8.0.24